+++ src/common/hooks/index.ts
@@ -1,4 +1,4 @@
 export { useTheme } from './DefaultUseTheme/DefaultUseTheme'
 export { useMediaQuery } from './DefaultMediaQuery/DefaultMediaQuery'
 export { useLearningPathTopic} from './DefaultLearningPathTopic/DefaultLearningPathTopic'
-export { useLearningPathTopicProgress } from './DefaultTopicProgress/DefaultTopicProgress'
+export { useLearningPathTopicProgress } from './DefaultTopicProgress/DefaultTopicProgress.hooks'
