+++ src/components/index.ts
@@ -14,6 +14,7 @@ export { default as MenuBar } from './MenuBar/MenuBar'
 export { default as ContactForm, type ContactFormProps } from './ContactForm/ContactForm'
 export { default as Footer } from './Footer/Footer'
 export { default as LocalNav } from './LocalNav/LocalNav'
+export { default as LocalNavItem } from './LocalNav/LocalNavItem'
 export { default as BreadcrumbsContainer } from './BreadcrumbsContainer/BreadcrumbsContainer'
 export { default as TextCardLeft } from './TextCard/TextCardLeft/TextCardLeft'
 export { default as TextCardRight } from './TextCard/TextCardRight/TextCardRight'
